fixed bug in DWORD64 utility function (thx theultramage for reporting) OFT: hopefully fixed problem with badly reported progress on transfer finished
git-svn-id: http://miranda.googlecode.com/svn/trunk@4377 83df23c2-51fd-11de-bb24-2d5e9b4af41d

+++ miranda/protocols/IcqOscarJ/icq_packet.c
@@ -547,7 +547,7 @@ void __fastcall unpackQWord(BYTE** pSource, DWORD64* qwDestination)
   if (qwDestination)
   {
     unpackDWord(pSource, &dwData);
-    *qwDestination = dwData << 32;
+    *qwDestination = ((DWORD64)dwData) << 32;
     unpackDWord(pSource, &dwData);
     *qwDestination |= dwData;
   }
